Handle non-catchable raids

+++ TeraFinder.Plugins/Utils/GridUtil.cs
@@ -18,6 +18,7 @@ public static class GridUtil
             { "GridUtil.Report", "" },
             { "GridUtil.RowsExceeded", "" },
             { "GridUtil.MismatchGroupID", "" },
+            { "GridUtil.NoCatch", "" },
         }.TranslateInnerStrings(language);
     }
 
@@ -160,43 +161,50 @@ public static class GridUtil
                     var encounters = f.Editor.GetCurrentEncounters(content, map);
                     if (EncounterRaidTF9.TryGenerateTeraDetails(seed, encounters, version, progress, eventProgress, content, map, id32, groupid, out var enc, out var result))
                     {
-                        var checkActiveHandler = ParseSettings.Settings.Handler.CheckActiveHandler;
-                        ParseSettings.Settings.Handler.CheckActiveHandler = false;
-                        if (!enc.GeneratePK9(result.Value, id32, version, f.Editor.SAV.OT, f.Editor.SAV.Language, f.Editor.SAV.Gender, out var pk9, out var la))
+                        if (enc.CanBeCaught)
                         {
-                            var la_encounter = la.Results.Where(l => l.Identifier is CheckIdentifier.Encounter).FirstOrDefault();
-                            if (content is RaidContent.Event or RaidContent.Event_Mighty)
-                                MessageBox.Show($"{strings["GridUtil.ErrorParsing"]}\n{strings["GridUtil.MissingData"]} [{enc.Identifier}].\n{strings["GridUtil.CheckWiki"]}");
-                            else
-                                MessageBox.Show($"{strings["GridUtil.ErrorParsing"]} {strings["GridUtil.Report"]}\n{la.Report()}");
-                            ParseSettings.Settings.Handler.CheckActiveHandler = checkActiveHandler;
-                            return;
-                        }
-
-                        var sfd = new SaveFileDialog
-                        {
-                            Filter = "PK9 (*.pk9)|*.pk9",
-                            FileName = $"{pk9!.FileName}",
-                        };
-
-                        if (sfd.ShowDialog() is DialogResult.OK)
-                        {
-                            if (File.Exists(sfd.FileName))
+                            var checkActiveHandler = ParseSettings.Settings.Handler.CheckActiveHandler;
+                            ParseSettings.Settings.Handler.CheckActiveHandler = false;
+                            if (!enc.GeneratePK9(result.Value, id32, version, f.Editor.SAV.OT, f.Editor.SAV.Language, f.Editor.SAV.Gender, out var pk9, out var la))
                             {
-                                try
-                                {
-                                    File.Delete(sfd.FileName);
-                                }
-                                catch (IOException ex)
-                                {
-                                    MessageBox.Show(ex.Message);
-                                }
+                                var la_encounter = la.Results.Where(l => l.Identifier is CheckIdentifier.Encounter).FirstOrDefault();
+                                if (content is RaidContent.Event or RaidContent.Event_Mighty)
+                                    MessageBox.Show($"{strings["GridUtil.ErrorParsing"]}\n{strings["GridUtil.MissingData"]} [{enc.Identifier}].\n{strings["GridUtil.CheckWiki"]}");
+                                else
+                                    MessageBox.Show($"{strings["GridUtil.ErrorParsing"]} {strings["GridUtil.Report"]}\n{la.Report()}");
+                                ParseSettings.Settings.Handler.CheckActiveHandler = checkActiveHandler;
+                                return;
                             }
-                            File.WriteAllBytes(sfd.FileName, pk9.Data);
-                            MessageBox.Show($"{strings["GridUtil.Exported"]} {sfd.FileName}");
-                        }
 
-                        ParseSettings.Settings.Handler.CheckActiveHandler = checkActiveHandler;
+                            var sfd = new SaveFileDialog
+                            {
+                                Filter = "PK9 (*.pk9)|*.pk9",
+                                FileName = $"{pk9!.FileName}",
+                            };
+
+                            if (sfd.ShowDialog() is DialogResult.OK)
+                            {
+                                if (File.Exists(sfd.FileName))
+                                {
+                                    try
+                                    {
+                                        File.Delete(sfd.FileName);
+                                    }
+                                    catch (IOException ex)
+                                    {
+                                        MessageBox.Show(ex.Message);
+                                    }
+                                }
+                                File.WriteAllBytes(sfd.FileName, pk9.Data);
+                                MessageBox.Show($"{strings["GridUtil.Exported"]} {sfd.FileName}");
+                            }
+
+                            ParseSettings.Settings.Handler.CheckActiveHandler = checkActiveHandler;
+                        }
+                        else
+                        {
+                            MessageBox.Show($"{strings["GridUtil.NoCatch"]}");
+                        }
                     }
                 }
                 catch (Exception ex)
@@ -380,20 +388,27 @@ public static class GridUtil
 
                     if (EncounterRaidTF9.TryGenerateTeraDetails(seed, encounters, version, progress, eventProgress, content, map, id32, groupid, out var enc, out var result))
                     {
-                        if (!enc.GeneratePK9(result.Value, id32, version, f.Editor.SAV.OT, f.Editor.SAV.Language, f.Editor.SAV.Gender, out var pk9, out var la))
+                        if (enc.CanBeCaught)
                         {
-                            var la_encounter = la.Results.Where(l => l.Identifier is CheckIdentifier.Encounter).FirstOrDefault();
-                            if (content is RaidContent.Event or RaidContent.Event_Mighty)
-                                MessageBox.Show($"{strings["GridUtil.ErrorParsing"]}\n{strings["GridUtil.MissingData"]} [{enc.Identifier}].\n{strings["GridUtil.CheckWiki"]}");
-                            else
-                                MessageBox.Show($"{strings["GridUtil.ErrorParsing"]} {strings["GridUtil.Report"]}\n{la.Report()}");
+                            if (!enc.GeneratePK9(result.Value, id32, version, f.Editor.SAV.OT, f.Editor.SAV.Language, f.Editor.SAV.Gender, out var pk9, out var la))
+                            {
+                                var la_encounter = la.Results.Where(l => l.Identifier is CheckIdentifier.Encounter).FirstOrDefault();
+                                if (content is RaidContent.Event or RaidContent.Event_Mighty)
+                                    MessageBox.Show($"{strings["GridUtil.ErrorParsing"]}\n{strings["GridUtil.MissingData"]} [{enc.Identifier}].\n{strings["GridUtil.CheckWiki"]}");
+                                else
+                                    MessageBox.Show($"{strings["GridUtil.ErrorParsing"]} {strings["GridUtil.Report"]}\n{la.Report()}");
+
+                                ParseSettings.Settings.Handler.CheckActiveHandler = checkActiveHandler;
+                                return;
+                            }
 
                             ParseSettings.Settings.Handler.CheckActiveHandler = checkActiveHandler;
-                            return;
+                            f.Editor.PKMEditor!.PopulateFields(pk9!, true);
                         }
-
-                        ParseSettings.Settings.Handler.CheckActiveHandler = checkActiveHandler;
-                        f.Editor.PKMEditor!.PopulateFields(pk9!, true);
+                        else
+                        {
+                            MessageBox.Show($"{strings["GridUtil.NoCatch"]}");
+                        }
                     }
                 }
                 catch (Exception ex)
